requirements.txt
fastapi
uvicorn[standard]
torch
diffusers
transformers
accelerate
opencv-python
imageio
safetensors
Pillow
